refactor: split CommandVisibility into domain-segregated modules (#468)
Split the 595 LOC CommandVisibility.ts "God Class" into 7 focused files:

- visibility/types.ts: CommandVisibilityContext interface
- visibility/helpers.ts: 12 shared utility functions
- visibility/TaskVisibilityRules.ts: 3 Task-specific functions
- visibility/ProjectVisibilityRules.ts: 4 Project-specific functions
- visibility/AreaVisibilityRules.ts: 2 Area-specific functions
- visibility/EffortVisibilityRules.ts: 12 Effort workflow functions
- visibility/AssetVisibilityRules.ts: 9 general Asset functions
- visibility/index.ts: Facade for backward compatibility

Benefits:
- Each file <100 LOC (SRP compliance)
- Domain cohesion: functions grouped by asset type
- Backward compatible: original import paths still work
- Zero logic changes: pure extraction refactoring

All 354 tests pass.
