+++ packages/core/src/domain/commands/CommandVisibility.ts
@@ -1,595 +1,15 @@
-import { WikiLinkHelpers } from "../../utilities/WikiLinkHelpers";
-import { AssetClass, EffortStatus } from "../constants";
-
 /**
- * Command Visibility Utilities
- * Contains pure functions for determining command availability based on asset context
- * Used by both CommandManager (for Command Palette) and React components (for buttons)
+ * Command Visibility
  *
- * Pattern: Strategy pattern for visibility conditions
- * Benefits: DRY, testable, consistent between UI and commands
+ * @deprecated Use domain/commands/visibility/* instead
+ * This file re-exports for backward compatibility.
+ *
+ * Prefer direct imports from domain-specific files:
+ *   import { canCreateTask } from "domain/commands/visibility/TaskVisibilityRules";
+ *   import { canCreateProject } from "domain/commands/visibility/ProjectVisibilityRules";
+ *   import { canCreateChildArea } from "domain/commands/visibility/AreaVisibilityRules";
+ *   import { canStartEffort } from "domain/commands/visibility/EffortVisibilityRules";
+ *   import { canCleanProperties } from "domain/commands/visibility/AssetVisibilityRules";
  */
 
-export interface CommandVisibilityContext {
-  instanceClass: string | string[] | null;
-  currentStatus: string | string[] | null;
-  metadata: Record<string, any>;
-  isArchived: boolean;
-  currentFolder: string;
-  expectedFolder: string | null;
-}
-
-/**
- * Check if instanceClass contains specific class
- */
-function hasClass(
-  instanceClass: string | string[] | null,
-  targetClass: string,
-): boolean {
-  if (!instanceClass) return false;
-
-  const classes = Array.isArray(instanceClass)
-    ? instanceClass
-    : [instanceClass];
-
-  return classes.some((cls) => WikiLinkHelpers.normalize(cls) === targetClass);
-}
-
-/**
- * Check if instanceClass is ems__Area or ems__Project
- */
-function isAreaOrProject(instanceClass: string | string[] | null): boolean {
-  return (
-    hasClass(instanceClass, AssetClass.AREA) ||
-    hasClass(instanceClass, AssetClass.PROJECT)
-  );
-}
-
-/**
- * Check if instanceClass is ems__Task, ems__Project, or ems__Meeting
- */
-function isEffort(instanceClass: string | string[] | null): boolean {
-  return (
-    hasClass(instanceClass, AssetClass.TASK) ||
-    hasClass(instanceClass, AssetClass.PROJECT) ||
-    hasClass(instanceClass, AssetClass.MEETING)
-  );
-}
-
-/**
- * Check if current status is Done
- */
-function hasStatus(
-  currentStatus: string | string[] | null,
-  targetStatus: string,
-): boolean {
-  if (!currentStatus) return false;
-
-  const statusValue = Array.isArray(currentStatus)
-    ? currentStatus[0]
-    : currentStatus;
-
-  if (!statusValue) return false;
-
-  const cleanStatus = WikiLinkHelpers.normalize(statusValue);
-  return cleanStatus === targetStatus;
-}
-
-/**
- * Check if asset is archived
- * Supports multiple formats: true, "true", "yes", 1
- */
-function isAssetArchived(isArchived: any): boolean {
-  if (isArchived === true || isArchived === 1) return true;
-  if (typeof isArchived === "string") {
-    const lowerValue = isArchived.toLowerCase();
-    return lowerValue === "true" || lowerValue === "yes";
-  }
-  return false;
-}
-
-/**
- * Check if metadata has empty properties
- */
-function hasEmptyProperties(metadata: Record<string, any>): boolean {
-  if (!metadata || Object.keys(metadata).length === 0) return false;
-
-  return Object.values(metadata).some((value) => {
-    if (value === null || value === undefined) return true;
-    if (typeof value === "string" && value.trim() === "") return true;
-    if (Array.isArray(value) && value.length === 0) return true;
-    if (
-      typeof value === "object" &&
-      !Array.isArray(value) &&
-      Object.keys(value).length === 0
-    )
-      return true;
-    return false;
-  });
-}
-
-/**
- * Check if folder needs repair
- */
-function needsFolderRepair(
-  currentFolder: string,
-  expectedFolder: string | null,
-): boolean {
-  if (!expectedFolder) return false;
-
-  // Normalize paths for comparison (remove trailing slashes)
-  const normalizedCurrent = currentFolder.replace(/\/$/, "");
-  const normalizedExpected = expectedFolder.replace(/\/$/, "");
-
-  return normalizedCurrent !== normalizedExpected;
-}
-
-// ============================================================================
-// Public API: Command Visibility Functions
-// ============================================================================
-
-/**
- * Can execute "Create Task" command
- * Available for: ems__Area and ems__Project assets
- */
-export function canCreateTask(context: CommandVisibilityContext): boolean {
-  return isAreaOrProject(context.instanceClass);
-}
-
-/**
- * Can execute "Create Project" command
- * Available for: ems__Area, ems__Initiative, and ems__Project assets
- */
-export function canCreateProject(context: CommandVisibilityContext): boolean {
-  return (
-    hasClass(context.instanceClass, AssetClass.AREA) ||
-    hasClass(context.instanceClass, AssetClass.INITIATIVE) ||
-    hasClass(context.instanceClass, AssetClass.PROJECT)
-  );
-}
-
-/**
- * Can execute "Create Event" command
- * Available for: ems__Area and ems__Project assets
- */
-export function canCreateEvent(context: CommandVisibilityContext): boolean {
-  return isAreaOrProject(context.instanceClass);
-}
-
-/**
- * Can execute "Create Child Area" command
- * Available for: ems__Area assets only
- */
-export function canCreateChildArea(context: CommandVisibilityContext): boolean {
-  return hasClass(context.instanceClass, AssetClass.AREA);
-}
-
-/**
- * Can execute "Create Instance" command
- * Available for: ems__TaskPrototype, ems__MeetingPrototype, and exo__EventPrototype assets
- */
-export function canCreateInstance(context: CommandVisibilityContext): boolean {
-  return (
-    hasClass(context.instanceClass, AssetClass.TASK_PROTOTYPE) ||
-    hasClass(context.instanceClass, AssetClass.MEETING_PROTOTYPE) ||
-    hasClass(context.instanceClass, AssetClass.EVENT_PROTOTYPE)
-  );
-}
-
-/**
- * Get today's date in YYYY-MM-DD format
- */
-function getTodayDateString(): string {
-  const now = new Date();
-  const year = now.getFullYear();
-  const month = String(now.getMonth() + 1).padStart(2, "0");
-  const day = String(now.getDate()).padStart(2, "0");
-  return `${year}-${month}-${day}`;
-}
-
-/**
- * Check if ems__Effort_plannedStartTimestamp is set to today's date
- * Handles timestamp format: YYYY-MM-DDTHH:MM:SS
- */
-function isPlannedForToday(metadata: Record<string, any>): boolean {
-  const plannedTimestamp = metadata.ems__Effort_plannedStartTimestamp;
-  if (!plannedTimestamp) return false;
-
-  const todayString = getTodayDateString();
-
-  // Handle string value
-  if (typeof plannedTimestamp === "string") {
-    // Extract date part from timestamp (YYYY-MM-DD from YYYY-MM-DDTHH:MM:SS)
-    const datePart = plannedTimestamp.split("T")[0];
-    return datePart === todayString;
-  }
-
-  // Handle array value (take first element)
-  if (Array.isArray(plannedTimestamp) && plannedTimestamp.length > 0) {
-    const datePart = String(plannedTimestamp[0]).split("T")[0];
-    return datePart === todayString;
-  }
-
-  return false;
-}
-
-/**
- * Can execute "Plan on today" command
- * Available for: Task and Project (any effort) that are NOT already planned for today
- */
-export function canPlanOnToday(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Hide button if already planned for today
-  if (isPlannedForToday(context.metadata)) return false;
-
-  return true;
-}
-
-/**
- * Can execute "Plan for Evening" command
- * Available for: Task or Meeting with Backlog status
- */
-export function canPlanForEvening(context: CommandVisibilityContext): boolean {
-  if (
-    !hasClass(context.instanceClass, AssetClass.TASK) &&
-    !hasClass(context.instanceClass, AssetClass.MEETING)
-  )
-    return false;
-
-  // Show only for Backlog status
-  return hasStatus(context.currentStatus, EffortStatus.BACKLOG);
-}
-
-/**
- * Check if ems__Effort_plannedStartTimestamp property exists
- */
-function hasPlannedStartTimestamp(metadata: Record<string, any>): boolean {
-  const plannedTimestamp = metadata.ems__Effort_plannedStartTimestamp;
-  if (!plannedTimestamp) return false;
-
-  if (typeof plannedTimestamp === "string") {
-    return plannedTimestamp.trim().length > 0;
-  }
-
-  if (Array.isArray(plannedTimestamp) && plannedTimestamp.length > 0) {
-    const value = String(plannedTimestamp[0]).trim();
-    return value.length > 0;
-  }
-
-  return false;
-}
-
-/**
- * Can execute "Shift Day Backward" command
- * Available for: Task and Project with ems__Effort_plannedStartTimestamp property
- */
-export function canShiftDayBackward(
-  context: CommandVisibilityContext,
-): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-  return hasPlannedStartTimestamp(context.metadata);
-}
-
-/**
- * Can execute "Shift Day Forward" command
- * Available for: Task and Project with ems__Effort_plannedStartTimestamp property
- */
-export function canShiftDayForward(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-  return hasPlannedStartTimestamp(context.metadata);
-}
-
-/**
- * Can execute "Set Draft Status" command
- * Available for: Task/Project without any status
- */
-export function canSetDraftStatus(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Show only when status is not set
-  return !context.currentStatus;
-}
-
-/**
- * Can execute "Move to Backlog" command
- * Available for: Task/Project with Draft status
- */
-export function canMoveToBacklog(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Show only for Draft status
-  return hasStatus(context.currentStatus, EffortStatus.DRAFT);
-}
-
-/**
- * Can execute "Move to Analysis" command
- * Available for: Project with Backlog status
- */
-export function canMoveToAnalysis(context: CommandVisibilityContext): boolean {
-  if (!hasClass(context.instanceClass, AssetClass.PROJECT)) return false;
-
-  // Show only for Backlog status
-  return hasStatus(context.currentStatus, EffortStatus.BACKLOG);
-}
-
-/**
- * Can execute "Move to ToDo" command
- * Available for: Project with Analysis status
- */
-export function canMoveToToDo(context: CommandVisibilityContext): boolean {
-  if (!hasClass(context.instanceClass, AssetClass.PROJECT)) return false;
-
-  // Show only for Analysis status
-  return hasStatus(context.currentStatus, EffortStatus.ANALYSIS);
-}
-
-/**
- * Can execute "Start Effort" command
- * Available for: Task or Meeting with Backlog status OR Project with ToDo status
- */
-export function canStartEffort(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Task and Meeting: Backlog → Doing
-  if (
-    hasClass(context.instanceClass, AssetClass.TASK) ||
-    hasClass(context.instanceClass, AssetClass.MEETING)
-  ) {
-    return hasStatus(context.currentStatus, EffortStatus.BACKLOG);
-  }
-
-  // Project: ToDo → Doing
-  if (hasClass(context.instanceClass, AssetClass.PROJECT)) {
-    return hasStatus(context.currentStatus, EffortStatus.TODO);
-  }
-
-  return false;
-}
-
-/**
- * Can execute "Mark as Done" command
- * Available for: Task/Project with Doing status
- */
-export function canMarkDone(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Show only for Doing status
-  return hasStatus(context.currentStatus, EffortStatus.DOING);
-}
-
-/**
- * Can execute "Trash" command
- * Available for: Task/Project without Trashed or Done status
- */
-export function canTrashEffort(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  // Show for efforts without status
-  if (!context.currentStatus) return true;
-
-  // Hide if status is Trashed or Done
-  const statuses = Array.isArray(context.currentStatus)
-    ? context.currentStatus
-    : [context.currentStatus];
-
-  const hasTrashedOrDone = statuses.some((status) => {
-    const cleanStatus = WikiLinkHelpers.normalize(status);
-    return (
-      cleanStatus === EffortStatus.TRASHED || cleanStatus === EffortStatus.DONE
-    );
-  });
-
-  return !hasTrashedOrDone;
-}
-
-/**
- * Can execute "Archive Task" command
- * Available for: Any asset that is not already archived
- */
-export function canArchiveTask(context: CommandVisibilityContext): boolean {
-  return !isAssetArchived(context.isArchived);
-}
-
-/**
- * Can execute "Clean Empty Properties" command
- * Available for: Any asset with empty properties
- */
-export function canCleanProperties(context: CommandVisibilityContext): boolean {
-  return hasEmptyProperties(context.metadata);
-}
-
-/**
- * Can execute "Repair Folder" command
- * Available for: Any asset in wrong folder (based on exo__Asset_isDefinedBy)
- */
-export function canRepairFolder(context: CommandVisibilityContext): boolean {
-  return needsFolderRepair(context.currentFolder, context.expectedFolder);
-}
-
-/**
- * Can execute "Rename to UID" command
- * Available for: Any asset where filename doesn't match exo__Asset_uid
- * Excluded: ims__Concept assets (concepts should keep their semantic names)
- */
-export function canRenameToUid(
-  context: CommandVisibilityContext,
-  currentFilename: string,
-): boolean {
-  const uid = context.metadata.exo__Asset_uid;
-  if (!uid) return false;
-
-  if (hasClass(context.instanceClass, AssetClass.CONCEPT)) return false;
-
-  return currentFilename !== uid;
-}
-
-/**
- * Can execute "Vote on Effort" command
- * Available for: Task and Project efforts (not archived)
- */
-export function canVoteOnEffort(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  if (isAssetArchived(context.isArchived)) return false;
-
-  return true;
-}
-
-/**
- * Can execute "Rollback Status" command
- * Available for: Efforts with non-null, non-Trashed status (workflow-based rollback)
- */
-export function canRollbackStatus(context: CommandVisibilityContext): boolean {
-  if (!isEffort(context.instanceClass)) return false;
-
-  if (isAssetArchived(context.isArchived)) return false;
-
-  if (!context.currentStatus) return false;
-
-  const statusValue = Array.isArray(context.currentStatus)
-    ? context.currentStatus[0]
-    : context.currentStatus;
-
-  if (!statusValue) return false;
-
-  const cleanStatus = WikiLinkHelpers.normalize(statusValue);
-
-  if (cleanStatus === EffortStatus.TRASHED) return false;
-
-  return true;
-}
-
-/**
- * Can execute "Create Related Task" command
- * Available for: ems__Task assets (not archived)
- */
-export function canCreateRelatedTask(
-  context: CommandVisibilityContext,
-): boolean {
-  if (!hasClass(context.instanceClass, AssetClass.TASK)) return false;
-
-  // Don't show button if archived
-  if (isAssetArchived(context.isArchived)) return false;
-
-  return true;
-}
-
-/**
- * Can execute "Set Active Focus" command
- * Available for: ems__Area assets only
- */
-export function canSetActiveFocus(context: CommandVisibilityContext): boolean {
-  return hasClass(context.instanceClass, AssetClass.AREA);
-}
-
-/**
- * Can execute "Copy Label to Aliases" command
- * Available for: Assets with exo__Asset_label that don't have this label in aliases yet
- */
-export function canCopyLabelToAliases(
-  context: CommandVisibilityContext,
-): boolean {
-  const label = context.metadata.exo__Asset_label;
-  if (!label || typeof label !== "string" || label.trim() === "") return false;
-
-  const trimmedLabel = label.trim();
-  const aliases = context.metadata.aliases;
-
-  if (!aliases) return true;
-
-  if (!Array.isArray(aliases)) return true;
-
-  if (aliases.length === 0) return true;
-
-  return !aliases.some((alias) => {
-    if (typeof alias !== "string") return false;
-    return alias.trim() === trimmedLabel;
-  });
-}
-
-/**
- * Can execute "Create Narrower Concept" command
- * Available for: ims__Concept assets
- */
-export function canCreateNarrowerConcept(
-  context: CommandVisibilityContext,
-): boolean {
-  return hasClass(context.instanceClass, AssetClass.CONCEPT);
-}
-
-/**
- * Can execute "Create Subclass" command
- * Available for: exo__Class assets
- */
-export function canCreateSubclass(
-  context: CommandVisibilityContext,
-): boolean {
-  return hasClass(context.instanceClass, "exo__Class");
-}
-
-/**
- * Can execute "Convert Task to Project" command
- * Available for: ems__Task assets only
- */
-export function canConvertTaskToProject(
-  context: CommandVisibilityContext,
-): boolean {
-  return hasClass(context.instanceClass, AssetClass.TASK);
-}
-
-/**
- * Can execute "Convert Project to Task" command
- * Available for: ems__Project assets only
- */
-export function canConvertProjectToTask(
-  context: CommandVisibilityContext,
-): boolean {
-  return hasClass(context.instanceClass, AssetClass.PROJECT);
-}
-
-/**
- * Extract date string from pn__DailyNote_day property
- * Handles wiki-link format: [[2025-11-11]] -> 2025-11-11
- */
-function extractDailyNoteDate(metadata: Record<string, any>): string | null {
-  const dayProperty = metadata.pn__DailyNote_day;
-  if (!dayProperty) return null;
-
-  if (typeof dayProperty === "string") {
-    const wikiLinkMatch = dayProperty.match(/\[\[(.+?)\]\]/);
-    return wikiLinkMatch ? wikiLinkMatch[1] : dayProperty;
-  }
-
-  if (Array.isArray(dayProperty) && dayProperty.length > 0) {
-    const firstValue = String(dayProperty[0]);
-    const wikiLinkMatch = firstValue.match(/\[\[(.+?)\]\]/);
-    return wikiLinkMatch ? wikiLinkMatch[1] : firstValue;
-  }
-
-  return null;
-}
-
-/**
- * Check if current date is greater than or equal to DailyNote date
- */
-function isCurrentDateGteDay(dailyNoteDate: string): boolean {
-  const today = getTodayDateString();
-  return today >= dailyNoteDate;
-}
-
-/**
- * Can execute "Create Task for DailyNote" command
- * Available for: pn__DailyNote assets when current date >= pn__DailyNote_day
- */
-export function canCreateTaskForDailyNote(
-  context: CommandVisibilityContext,
-): boolean {
-  if (!hasClass(context.instanceClass, AssetClass.DAILY_NOTE)) return false;
-  if (context.isArchived) return false;
-
-  const dailyNoteDate = extractDailyNoteDate(context.metadata);
-  if (!dailyNoteDate) return false;
-
-  return isCurrentDateGteDay(dailyNoteDate);
-}
+export * from "./visibility";
